refactor to cover check styles
