+++ src/integrationTest/java/uk/gov/hmcts/reform/civil/config/FeignRequestUtils.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.controllers;
+package uk.gov.hmcts.reform.civil.config;
 
 import feign.Request;
 import org.springframework.http.HttpHeaders;
@@ -8,24 +8,16 @@ import java.util.ArrayList;
 
 public class FeignRequestUtils {
 
-    /**
-     * Convert Feign {@link Request} headers collection to Spring {@link HttpHeaders} object
-     *
-     * @param request Feign request to convert
-     * @return Converted HTTP headers (as a Spring object)
-     */
+    private FeignRequestUtils() {
+        //Utility class
+    }
+
     public static HttpHeaders convertRequestHeaders(final Request request) {
         final HttpHeaders springHeaders = new HttpHeaders();
         request.headers().forEach((header, vals) -> springHeaders.put(header, new ArrayList<String>(vals)));
         return springHeaders;
     }
 
-    /**
-     * Convert Feign {@link Request} request method to Spring {@link HttpMethod} object
-     *
-     * @param request Feign request to convert
-     * @return Converted HTTP Method (as a Spring object)
-     */
     public static HttpMethod convertRequestMethod(final Request request) {
         return HttpMethod.valueOf(request.method());
     }
